Added Embedding support for Web safe Audio and Video formats.
The changes in this commit make it so that the HTML writer embeds some web safe audio and video files.

In addition to that the ".webp" file format is now also embedded, which is on par with the default behaviour of Discord.

The new supported formats are:

Image:
    .WEBP

Audio:
    .MP3
    .WAV
    .OGG
    .FLAC
    .M4A

Video:
    .MP4
    .WEBM